This is what I expected the allocation-numbering bit to look like.
We clear out caching code factories aggressively to avoid huge memory
bloat, so you simply *can not* dump a code factory way after its been
used and mutated and expect it to still contain representations for
your code.  I'm about to check in a hack to SAMain to show how this
numbering object should be used.