CygWin portability fixes for GNUmakefile, Loader and build-precisec: translate CLASSPATH to windows format if on CYGWIN via the cygpath utility and add the CLASSPATH to the java/jikes command line explicitly. Also, the Windows port of the JDK seems to have broken .zip file handling: paths in the ZIP file are only found if they use '/' to delimit paths, instead of the architecture-specific path-delimiter character.  We hack around this brokenness in ClassFile.Loader.
Finally, two unix portability fixes: openssh doesn't have the -A option
which we were specifying in the makefile (we now use -q instead), and the
Harpoon.jar target was breaking with 'too many arguments' on some platforms.
We hacked around this last problem by using the -u option to jar [which
is only present in JDK1.2 and above =( ] to build the jar file incrementally,
package-by-package.