SyncTransformer bugfixes:
checkForAbort should only happen when we've got a method-local transaction
context.  Other times the abort exception should just get thrown up a level.

We don't *fail* when we encounter a synchronized reference to a static field
anymore, as they seem to be fairly common, though harmless (mostly confined
to static initializers which are invoked in a single-threaded context
anyway).  We print a warning message instead.

Likewise, in addChecks() we don't fail if we're invoked on a PHI: we just
assert() that no checks are actually being requested for this location
before returning.

makeFlagConst() and its associated flag values are being moved out of
the inner class because javac (although not jikes) complains about
*any* static members of inner classes.  (Jikes just says the static
fields must be final.)