Updated & debugged DispatchTreeTransformation.  It works now, at least for the examples I've tried.
Changed the infrastructure to support input and analysis in SSI form,
and mutation and output in RSSx form.  We use a proxy map for the
type info and etc that applies to the SSI form input.

We now gauge suitability for CALL replacement based on the number
of *non-abstract* children, since abstract children can never be
instantiated and thus don't need to be tested against.

The actual CALL replacement now uses ClassHierarchy.overrides() instead
of its own implementation. This reduces code duplication and probably
also fixes the subtle interface method inheritance issues which
ClassHierarchy.overrides() takes pains to get right.  We now filter
the method list to remove uncallable, abstract, and interface methods,
and sort from most-specific to least-specific externally.  We're also
careful to skip call sites which are unreachable (have no valid targets)
so that we don't generate linker references to uncallable methods.