From examination of the StrongARM assembly output for the _222_mpegaudio SPEC benchmark, it was discovered that patterns of the form CJUMP(BINOP(cmpgt, CONST(c), r)) were causing code similar to   mov r0, #c   cmp r0, rX to be emitted, instead of the desired:   cmp rX, #c
This turned out to have two causes: first, AlgebraicSimplification was
not moving the constant to the right on <, <=, >=, and > tests as it
does with commutative binops.  Also, the CommutativityExpander was not
expanding the existing CJUMP(BINOP(cmpop, r, CONST(c))) pattern in such
a way as to match trees where the CONST was on the left.  Both these
oversights have been corrected:

 - Analysis/Tree/AlgebraicSimplification now moves the constant in
   a constant comparison to the right.  The new rule is called 'commute2'.
 - CommutativityExpander properly expands cmpop CONST patterns even
   when the cmpop is not "completely" commutative -- ie, the cmpop
   is CMPLT, CMPLE, CMPGE, or CMPGT, which means that the cmpop needs
   to be flipped when the arguments are.  This is done with a fancy
   extra predicate in the tricky cases, which must appear first in
   the predicate list.