Completed implementation of intelligent FieldOracle and CheckOracle.
The HoistingCheckOracle attempts to hoist and coalesce checks as far up the
dominator tree as possible (the CheckOracle abstract class was patched
to make CheckOracle.RefAndField and CheckOracle.RefAndIndexAndType
collections-friendly as part of this).

The GlobalFieldOracle does a whole-program analysis to determine which
are the fields that can possibly be accessed in synchronized and
unsynchronized contexts.  More relevantly, you can also determine
from this which fields *can't* be accessed in those contexts.  This
allows us to skip some checks occasionally in the generated transactional
code.