Changes to Runtime to accomodate RTJ.
Run ./setup to process the new configure.in, and
use the --with-realtime-java option to configure with RTJ extensions.

Note that if you ./configure --with-realtime-java ..., then you will
have to run bin/build-precisec with the -t option to enable Realtime Java
support.  The Runtime and compiled output must either both have RTJ enabled
or disabled in order to produce a working executable.  This is explained
more fully in ~/Harpoon/Realtime/BUILDING.

jni-private.h's aarray was modified to add padding for the memoryArea field.
java_lang_Thread initializes the runtime with a RealtimeThread (in case
WITH_REALTIME_JAVA is turned on...)