Split Realtime (which was getting too big) into ThreadToRealtimeThread, CheckAdder, and Realtime.
QuadCounter and Timer are two utility classes that you might find useful.

QuadCounter is an HCodeFactory that simply counts the number of Quads that
are converted.

Timer is a class which can time blocks of code in the compiler.  Use
timer.start() and timer.stop() around blocks of code and timer.timeElapsed()
to get the cumulative amount of time elapsed between starts and stops thus
far.  For example:

Timer timer = new Timer();
for (int i=0; i<10; i++) {
  bar();
  timer.start();
  foo(i);
  timer.stop();
  baz();
}

timer.timeElapsed() returns the total amount of time spent in foo()
(not bar() or baz()).