Improved SCC analysis: we now propagate lattice values of the form "the result of an INSTANCEOF" and "the result of a boolean OPER" through the algorithm, which allows us to refine types at SIGMAs *much* more aggressively.  In particular, we should optimize out more typechecks in the common idiom:
   if (x instanceof Foo) {
    Foo f = (Foo) x;
   }

Before, the checks were being left in because the low-level arcana of
quad translation was inserting a PHI in just the wrong place, and SCC
analysis couldn't see past the PHI.  Now we can always see past PHIs
and SIGMAs to do the optimization.  Yay, team.

Also added an 'isPossiblyNull()' accessor method to SCCAnalysis which
allows better interrogation of the results.  This is mostly for
debugging, but who knows, maybe it will be generally useful.