1. The MetaCallGraph code was improved and cleaned.
2. The bugs introduced in 1 were removed. :-)

3. A new flag (RECORD_ACTIONS) controls whether we
record the actions done by the analyzed program or
not (default is false). This info seems to be very
expensive to collect and not that useful in practice.

4. PAMain: A new flag (check_nc) lists all the call
sites with 0 callees. This happens in general for
the calls on SecurityManager objects as we don't
instantiate any of them (e.g SecurityManager.checkWrite
etc). It doesn't necessarily correpond to error in our
(meta)call graph construction, but to unreachable portions
of code. It also prints some statistics on the number of
virtual/non-virtual call sites. After initial tests, it
seems that the overwhelming majority of the call sites
can be devirtualized (I have no idea on how much does
this affect the dynamic behavior of the program).

5. Various small changes/improvements.