Add preliminary support for alternate array implementations.
This includes "generic java type" support, where we can generate 9
different "specialized" versions of a template, one for each primitive
java type and Object.  The generic file has a filename ending in
JavaType.jt (and should be specified in MACHINE_SRC and MACHINE_GEN in
GNUmakefile); from this the nine files are generated by substituting a
captialized type name for every instance of the string "JavaType"
(including in the filename) and the appropriate 'real' type name
(lowercased except in the case of Object) for every instance of the
string "javaType".

How I wish Java had real parameterized types.  This is a crude workaround.
(Alternatively, Java could have fully first-class primitive types, which
wouldn't require such evilness).