The MAInfo module has been modified to support multi-level inlining to increase the effectiveness of the stack allocations.
In short: suppose a method C allocates some objects
(nodes in our abstraction) that escapes only into the caller
of C (let's call it B) and from B into its caller (A).
Inlining the call chain C, B, A into A will allow for those
objects to be stack allocatable.

In practice, there are complications due to the fact that
different inlining chains might have common subparts and
so, the order of inlining becomes VERY important. The current
implementation uses the optimal ordering.

For the time being, this is part of MAInfo. It will be interesting
to implement it as a separate class, thus allowing other parts of
the compiler to use it.