javadoc updates to Runtime/Counters.java, Analysis/Counters/package.html, and Analysis/Counters/CounterFactory.java.
Added codeFactory() to CounterFactory (and EpilogMutator class in
harpoon.Analysis.Counters which supports it) to add calls to
harpoon.Runtime.Counters.report() before any call to Runtime.exit()
or any RETURN or THROW in the main method.

Fixed a bug in spliceIncrement() that was locking the *counter* instead
of the *lock*.  Oops. =)