Removed an extraneous comma that caused a ",: command not found" error when configuring.
Generalized the 'object padding' support to make it easier for
transformations to add fields to java.lang.Object (and thus to arrays as
well).  Now --with-object-padding=N builds the runtime with support for
N extra bytes of fields in java.lang.Object.  --with-realtime-java
adds 4 to object padding by default, as the realtime transformation adds
one 4-byte field to java.lang.Object.

Incidentally, if you use the useUniqueRWCounters option in SyncTransformer,
you should build the runtime with --with-object-padding=8.