Add k-groups to dataflow lattice.  The analysis now correctly groups together (for example) memory accesses to <basepointer>+(0 mod N) and <basepointer>+(1 mod N), if gcd(N,32)>1.
This detects cache-equivalence groups in manually unrolled loops.
When we write a loop-unrolling pass, it will detect eq groups there,
too.

Currently, the k-group definition is a little funky, so that the first
iteration of an unrolled loop is not correctly put in the same k-group
as the other iterations.  Fixing this is not difficult (basically,
k-groups should represent phi functions, ie definition *groups*,
instead of single definition *points*) but I'm going to commit what
I have first, since it works. =)