LoopFinder wasn't properly initializing 'grapher' in the private constructor, causing a null pointer error.
Changed return type of loopBackedges() and loopExits() to Set of
HCodeEdges, instead of Set of HCodeElements.  The edge information
is much more useful in this case.  Changed javadoc in
Analysis/Loops/Loops.java to reflect this, too.

Oddly enough, no other analysis in the infrastructure currently uses
these two functions.