fixed a big thinko: i was using the size of objects with a given type to enable more agressive small-object optimization, without realizing that *objects can be subtyped* and so can have a larger size than their type would at first seem to indicate.  We now look at all of a types subclasses and test against the maximum before enabling the small-object optimization.
Although this change is necessary for correctness, its impact on
statistics should be small.  Emmett only found one violation in SPEC
(in Jack) which could be attributed to this bug, and there only
because the small-object optimization was taking precedence over the
more exact analysis which would have properly separated cache
equivalence classes.  Scott's claims:
 a) class hierarchies are generally broad, not deep.
 b) I suspect even deep subclasses are generally pretty small.
 c) I expect that most local variables (the only types we look at,
    at present) are typed rather closely.

By tomorrow Emmett should be able to tell me if these assumptions are
correct, or whether our cache-equivalence statistics are changed
drastically by this bug-fix.