Hack for Wes' RTJ stuff and continuing refactoring of pointer-reversal algorithm.
The GCs currently do not handle arrays properly when Objects contain
fields, which is the case for Wes' and Scott's stuff. Current hack
assumes that these fields are all pointers, which should work for
Wes but is not a good long-term solution. Will fix once the pointer-
reversal stuff is under control.