Added 'updateKey()' method to Heap interface.  This usually has poorer performance guarantees than 'decreaseKey()' and is almost (but not quite exactly) the same as a 'delete()' followed by an 'insert()' with the new key value (delete-than-insert would change the Map.Entry associated with the key-value pair, while 'updateKey()' maintains it).
Refactored some code in AbstractHeap to make this easier to implement,
without undue code duplication.  Also added testcases for the updateKey()
method to the various heap implementations.

It is possible to slightly-more-efficiently implement 'updateKey()' by
changing the BinomialHeap and FibonacciHeap data structures slightly.
I believe the time-bounds change to requiring O(lg n) time for m
calls to 'updateKey()' instead of O(m lg n).  Implementation is
deferred until it can be shown truely necessary.  See the discussion at
the end of section 3 of "A Data Structure for Manipulating Priority
Queues" by Jean Vuillemin in Communications of the ACM 21(4) for the
change to BinomialHeap; and the "good-tree/bad-tree" variant of
FibonacciHeap discussed in section 3 of "Fibonacci Heaps and Their
Uses in Improved Network Optimization Algorithms" by Fredman and
Tarjan in Journal of the ACM 34(3).

A trivial modification applies to BinaryHeap to make updateKey
more efficient (calling downHeap rather than upHeap if the
key is increasing); I will implement this in the next commit.