Add support for masked pointers to Runtime1.  There's a hack in DerivationGenerator to make it give proper derivations after a pointer AND, but it's a little overbroad, and will give derivations even when the AND completely obliterates the original pointer.  It is currently the responsibility of the implementor to DECLARE a proper type for such constructions, to override the DerivationGenerator's guess. At some future point I may make the DerivationGenerator smarter about which pointer ANDs it generates derivations for, but not at the present moment (I'm about to leave the state for the weekend---ah, summertime!).
There's also a loose end with respect to pointer comparisons.  The
code currently assumes that any twiddled low bits of the pointer are
integral to it; it will not compare as equal with a pointer with
different low bits.  Sometimes you may want to mask the low bits
before doing the comparison.  At the time I wrote
Generic.Runtime.TreeBuilder I never thought that *pointer-comparison*
would be potentially runtime-dependent.  I will probably eventually
add another method to the generic TreeBuilder to allow the Runtime to
specify proper handling of the compare operation.

The Runtime1.TreeBuilder constructor could use some prettify-ing,
too, but I want to make sure it's (eventually) possible to specify
masking both programmatically (enabling realtime java also
automagically enables pointer masking, for example) and from the
command-line (-Dharpoon.runtime1.pointer_masking=true, or some such).
Not done yet, but will be done after the weekend.

Now it's time for people to bang on this code with pointer masking
enabled and allocators that (say) set the low bit, and see if I
missed any masking cases.  I don't think so, but I'm not infallible.