Fixed NoHeapRealtimeThread support - now works.  Now, native methods also go through RTJ_malloc when allocating objects (which could potentially be returned to Java).  Added some more options for RTJ in configure.in, now everything can be automated by scripts.
There are still some minor bugs left in the RTJ system - fake scopes
doesn't quite work yet, natively-allocated objects may not be tagged
correctly (right now, they're tagged in such a way that all access checks
will pass, which just eliminated error-checking for native methods -
may fix in the future).  If you enter an LTMemory, then enter a HeapMemory,
then enter an LTMemory and start a NoHeapRealtimeThread, if that thread
attempts to make a reference from the current scope to the first LTMemory,
the system will fail with a heapCheck error (provided heapChecks are turned
on).  Will fix this in the future too, but since none of my benchmarks do
this, it's not critical.

Static initializers attempting to allocate from a MemoryArea other than the
heap may cause infinite loops if they're located in a Thread object or any
object called during setup of the main thread - will punt this forever,
since it only concerns a small number of javax.realtime.* classes and those
classes which inherit from javax.realtime.* classes - and it seems pretty
unavoidable, the system produces useful error messages in this case,
and there are simple work-arounds.

The GarbageCollector, PhysicalMemory, RawMemoryAccess, and RealtimeSystem/
security classes are not finished yet, but I'm punting that for now as
well until the thesis is written.

The system appears to be stable enough for significant benchmarking and
writing of thesis and papers...  It's also stable enough for Brian to
integrate his Realtime thread stuff with it.