Lots of fixes to SCCAnalysis. Revamped the value lattice so that xOperBooleanResult and xInstanceofResult extend *both* xBitWidth *and* xIntConstant (depending on whether the result of the comparison is known). This (along with some changes in handleSigmas) ensures that the information propagated from the CJMP source makes its way into the sigma functions regardless of whether the destination of the jump is statically known or not.  Refactored the "no information known" handleSigmas function into its own method, too, for more uniform handling of the branch condition.  Unified the 'left is source' and 'right is source' legs of the handleSigmas(xOperBooleanResult) method, and fixed some bugs in the handling of ICMPGT/LCMPGT.
Also added a whole bunch of assertions to visit() methods to ensure
that the static properties of the quads (object references not null,
etc) are in fact being properly analyzed.  This assertions now hold
for *almost* the entire SPEC suite: in _228_jack we throw an assertion
failure because UnHandler is smarter than SCCAnalysis: in particular,
UnHandler detects an aliasing relation between stk0 and lv1 which
means that a null-check on lv1 protects stk0 as well.  SCCAnalysis
can't figure this out yet.  After lunch I'll work on trying to
make SCCAnalysis smart enough to figure this out: SSI form handles
*most* of these aliasing questions by only having one temp per value;
but here's obviously a case where that rough generalization breaks
down.

Oh -- and all of SPEC (with the exception of _228_jack, discussed
above) executes correctly with SCC optimization turned on, so this
code has a very good chance of being correct.