Changed GCstats so that will work without RTJ.  Unfortunately, BDW was leaking like a freak because GC_register_finalizer was not being called when RTJ was turned on.
This also brings up the issue that we may not be doing finalization
in precise_gc.  Karen - please update jni-inflate to include a call to your
GC finalization function so that we can deflate the many inflated objects that
can created by Barnes and Water (to provide synchronization for sync. removal).