Bug fix to Peephole: the SwapVisitor was ignoring handler status, resulting in swaps of protected and unprotected quads.  This (along with the bug described in the previous commit to this file) caused problems compiling harpoon.IR.Quads.RSSxToNoSSA, because the definition of lv2_0 was being moved across the handler boundary. This caused potential executions of the handler before lv2_0 had been defined, resulting in type errors converting to low-quad-ssi form.
The Peephole visitor has been revamped, too.  Instead of forbidding
any change to 'protected quads' (those with handlers), we use a much
simpler rule: no movement of quads is allowed across handler boundaries.
[We also had to reorder the 'Qp instanceof FOOTER' rule, because the
FOOTER was being interpreted as a new handler context.  Now we deal
with FOOTERs first.]  This rule should be applicable for all well-formed
code generated from java source, as no variables defined inside the try
block may legitimately be referenced from outside the try or in the
catch or finally, so moving definitions around inside the try block
is acceptable.  I *think* the rule is applicable for all bytecode
which passes the verifier, too, but I'm not certain.  [As part of
these new simpler rules, we re-enable moves past calls when the CALL
is in the same handler context as the MOVE.]

The 'allowFarMoves' parameter is now unused.  The rules described above
work perfectly well for both quad-with-try and no-try forms.

Finally, we use a UniqueStack instead of a simple Stack now for the
to-do list, which results in less needlessly duplicated work.