Bug-fixes for ToNoSSA: I'm not sure how this class ever worked.  Perhaps it was never really used?
ToNoSSA was trying to merge temps *and* clone quads at the same time--bad idea!
It would often discover a necessary merge *after* it had already cloned
several quads with the *old* name of the Temp, and never go back and
update them with the merged name.  The fix is very simple: we go through
the code first quickly and do all the merges, *then* we go back and
do all the cloning.  The functions of the NameMap and the CloningTempMap
have been merged, with a boolean and assertions to keep the merging
step separate from the cloning step.