Added 'MapTest', which is test code (liberally borrowed from the Mauve project) to test implementations of java.util.Map for correctness. Note that because our MultiMap interface is a semantically consistent extension of Map (ie, all the Map methods behave properly if you don't use any of the MultiMap features), we can also use MapTest to (partially) test the MultiMap implementations for correctness.
java.util.HashMap, GenericInvertibleMap, GenericInvertibleMultiMap,
and GenericMultiMap pass this test suite with no errors.

harpoon.Util.LinearMap fails in the entrySet() method checks, because
it doesn't prohibit you from 'add'ing elements which are not Map.Entry's
to the entrySet.  This is a safety error but not a correctness problem.

harpoon.Util.HashEnvironment doesn't support the remove() operation on
its entrySet(), keySet(), or values() collections; the test-suite
reports this failure.  Otherwise it passes.