Improvements/bug-fixes to DefiniteInitOracle: bug-fix in findThisVars that was causing the set of thisvars to always be empty -- basically we were treating the definition of the canonical 'this' in the METHOD quad as a destructive definition of a not-this variable.  Oops.  Rearranged the iteration code at the end so that it's a while instead of a do-while loop.  This was just a clean-up.
Improvements to findNextExitPoint: THROWs are not exit points, since
the constructor does not complete successfully.  Neither are CALLs to
'this' or 'super' constructors.

Improvements to findDefInit: a call to a 'this' constructor is a
definite initialization of all the fields which that constructor
definitely initializes (recursive invocation of the analysis).
We changed factored out a 'getNotDefInit' method and had it
cache its results to make the recursive invocation work best.