The evil optimizations to make the benchmarks work out nicely begin.
This check-in adds a gnarly special case to the MayReadOracle to help
improve the accuracy of DefiniteInitOracle.  Basically, you can have
constructors of the form: (this example taken from _202_jess)
  Token(Token t) {
    this.tag = t.tag;
    ...
  }
The MayReadOracle normally says that field 'Token.tag' is "possibly read"
before the write to 'this.tag'.  However, in this case we know that
't.tag' is not a write to 'this.tag' because no parameters to a constructor
can be 'this' (or an alias) (excluding the very first, hidden, parameter,
which is *always* 'this').  So now MayReadOracle has a
'specialHandlingForConstructors' mode where it ignores writes to fields
whose object ref *must* be a non-'this' parameter.  This mode probably
isn't useful for anything but DefiniteInitOracle --- but that's alright,
turn it off if you don't need it.