Extended the runtime selection to use reflection.  This way we don't have to keep changing this code whenever we add a new runtime subclass. Kept a static reference to Runtime1 so that FLEX built w/ itself doesn't need a manual root-set specification.
I'm not entirely sure I'm happy w/ the cleanness of this code, so I'm not
applying the change to all the Backends' Frames yet.  I'm going to let it
sit for a while in the tree, and if it doesn't make me too unhappy, then
I'll spread it to the rest.