Added runtime support for claz shrinking.  Abstracted the 'claz' field of struct oobj* so that you now use FNI_CLAZ(struct oobj* someobj) to get at it; this way the code "does the right thing" whatever size and encoding the claz field happens to take.
Code in Karen's precise GC is too tightly tied to the claz field to
make this work; as an example, the forwarding pointer is stored in
the claz field, which doesn't really work if I shrink the claz field
down to a handful of bits.  This isn't really a problem: we just don't
use precise garbage collection in conjunction w/ claz shrinkage.
The build will fail horribly if you try to configure --with-claz-shrink
and --with-gc=precise.  So don't do that.

Note that claz shrinkage only occurs if you configure --with-claz-shrink
and that your Code must be similarly configured (use the
-Dharpoon.runtime=Tiny option).  The configuration checking will make
your final link fail unless your Runtime and Code match.