GJ-ized harpoon.Util.FilterIterator and harpoon.Util.Collections.AggregateSetFactory.
Note that: 1) the jsr14 GJ spec doc has a much more elegant means of doing
mapping functions, using non-static inner classes.  See the 'Zipper'
example in the text.  We used an more inelegant approach to maintain
compatibility w/ existing code.  Maybe someday we'll "upgrade".
2) AggregateSetFactory is one of the places where one would wish that
the types of the generated sets we're so tightly tied to the type
of the Factory; i.e. it would be nice to have a generic factory that
could make Set<V>s for any V.  However, the MapSetFactory (which turns
MapFactorys into SetFactories that generate MapSets) would then be
impossible to type-check.  Ce la vi.