Work around javac bugs.
In the CFGrapher case, the code was valid but yielded an 'unchecked'
warning; but javac massacred the code-gen resulting in a VerifyError
thrown when the bytecode for the class was executed.  Corrected this
by fixing the decl to not generate the 'unchecked' warning.

In the UnmodifiableListIterator case, this is bug in how the GJ compiler
handles final methods.  I've reported it to Sun and hopefully it will
be fixed.  In the meantime, we must mark the methods as non-final to
work around the problem.