Keep up with GJ prototype 1.2 changes: new signature for addAll, putAll, etc and constructors are covariant.
harpoon.Util.Default has new unmodifiableCollection() method which doesn't
yet work properly due to a bug in the GJ compiler; we've worked around
the bug by adding an extra parameter to help the type inference engine.
We also rejiggered the type signature of the Default.pair() method to
allow pair elements of differing types so that we don't break as much
existing code.

We also narrowed the type of MultiMap.entrySet() to require a MultiMapSet
as the return value.