1. Modified the constructors for meta / smart call graphs: they now require a linker (in addition to their former parameters). All classes that called this constructors have been modified too.
2. Found and fixed a (quite important) bug in MetaCallGraphAbstr.java
It's funny that it never occured in practice before!  Isn't this
amazing?  Maybe someone up there really cares about me :)

3. Added some code top the meta call graph to keep track of the things
that are put in collections and maps.  This allows us to trim the call
graph a bit.  More can be done but I think the real solution is to
have a real Data Polymorphic CPA-like call graph construction.