Speed up compile times by only rebuilding source files which have changed since the last compilation.  Deleting the file 'out-of-date' will force everything to be rebuilt.  Also, if you type 'make' when *nothing* seems to be out-of-date, we'll rebuild everything, just to be sure.
Let me know if this makefile change behaves as you expect or if you find
files which for some reason or another are not being properly rebuilt or
are always rebuilt or some such.

With this change to the makefile, build times drop (for me) from  52s user
time to under 9s (i.e > 5x speedup).  If you're not using JDK1.4 already,
you might consider grabbing that, too, as the javac in 1.4 is pretty fast.
Hope this makes your compiling experiences happier!