Re-merged cata's changes to startup.c.  This is how all these patches should have looked. *PLEASE PLEASE PLEASE* don't commit any file with crazy 'every line is different' diffs because of weird whitespace conversions! If 'cvs diff -u' doesn't look clean and simple, don't commit.
Compare revisions 1.32, 1.33, and 1.35 of this file to see how this commit
*should* have looked.

This compiled and should work for non-realtime threads.  No guarantees
for the realtime threading case.  Compare this version to the diff
between 1.32 and 1.33 (use the '-b' option to diff to suppress the
completely bogus whitespace changes) to see how I reworked it if things
break for you in the realtime thread case.