FNI_GetClassInfo changed from FNI_GetClassInfo(jobject) to FNI_GetClassInfo(env, jobject)
FNI_WRAP requires an env to be declared in the scope

RTJ_MALLOC_UNCOLLECTABLE expanded to malloc and for some reason,
gcc warned me about returning a void* from malloc - which is
exactly what malloc returns.  A cast to void* eliminates the bogus
warning - and void* is exactly what a general-purpose allocator like
allocfunc should return.