Removed Main.java - leftover debugging code from Cata's debugging sessions.  Please don't check in unconditional debugging output commands... In RTJ, you should use NoHeapRealtimeThread.print (which is non-portable) because the RTJ implementation may or may not have access to System.out depending on whether you're in a NoHeapRealtimeThread or if static initializers have run...
We currently do not support Thread.start or MemoryArea.enter in static
initializers - correct behavior for these should be to block until
static initialization has completed.  This might cause a deadlock
unless the static initialization completes anyway and the work is
rescheduled for after static initialization (after the Scheduler has
started but before the main thread is run).  Care must be utilized
to ensure this does not cause cycles in the static initializers.

Also fixed getDefaultScheduler() - note this method is explicitly
not thread-safe, but it becomes thread-safe after first run.
getDefaultScheduler() is run right after static initialization, but
before starting the main thread.  Therefore, considering the larger
context, it is thread-safe.  This is the case with most RTJ startup code.
Since thread.start would delay after RTJ is up and running - even if
we handle thread.start correctly during static initialization, this
will not cause the RTJ methods to become non-thread-safe.