So the piece of code demarked: "by cata: The dies horribly, and I'm too lazy to debug it..." dies because he frees memory and leaves a GlobalRef dangling. If he had GC enabled, GC would follow the ref into freed space looking for roots.
Also, the reason he "didn't do exit" is because his thread scheduler
depends on the fact that when there's no work left to do, the
currentThread gets set to NULL - so everything can drop through.

The problem is that currentThread->jnienv is the only access to env
and precisec generated C code depends on the env being available to keep
GC roots updated with env->local_refs.  Thus, once the main thread
has been punted, there is no currentThread to lookup the env to run
Thread.exit() (and keep GC local_refs updated).

The real answer to this is to rewrite his entire scheduler so that
one thread remains at the end of main - not no threads, but that
the scheduler no longer has to concern itself with running that thread once
the main method finishes.

Thread.exit sees a half broken world in most implementations of Java VM's.
We just gotta make sure it isn't TOO broken.  Therefore, the setupEnv(env)
method sets up a world that's good enough to limp through Thread.exit()
and die at destroyEnv()