Added lots of debugging info enabled by RTJ_DEBUG_THREADS... ... and found a bug that prevented threads from being run.
addToFeasibility is the method that adds a thread to a thread list.
It is called on the Scheduler that the thread is attached to
(Thread.getScheduler()).  If a thread is not attached to a Scheduler
(null by default), then the thread is not run when the start method is
called (which makes sense).  Thus, RealtimeThreads must call setup in
the constructor, which then sets the currentScheduler for the thread
to the Scheduler.getDefaultScheduler() (initially a PriorityScheduler)
so that they can be added to the thread list in the addToFeasibility call
on RealtimeThread.start().  If you want to set the scheduler, that's
fine - but it better start out as PriorityScheduler (or some scheduler other
than null).  The insight I gained from the debugging info. is that setting
the scheduler is NOT the business of the runtime (except for the main thread,
which cannot set it initially in the setup because RTJ_init hasn't happened
yet - which means that ImmortalMemory.instance() cannot be called to get
the instance from which the PriorityScheduler must be allocated (to allow
NoHeapRealtimeThreads/scope-allocated threads to touch the scheduler
to schedule themselves).  Since entering ImmortalMemory.instance() relies on
the RealtimeThread.currentRealtimeThread()->inflated_obj->memBlock to be
set properly before allocation in the memory area can begin, the main
thread is in the wierd circumstance that it must be scheduled by a
PriorityScheduler that can't be allocated until the thread is running.
That's why the main thread must have its scheduler set by runtime code
after the execution of RTJ_init.  But even if the defaultScheduler must
be set to null (to prevent a cycle in the main thread) by default, if
RTJ_init is finished (which happens before any thread but the main thread
is run) - the defaultScheduler must then be set to the proper (non-null)
PriorityScheduler to enable scheduling of threads other than the main thread.

Whew!