Sort the list of prototypes at the start of the c file so that output is repeatable across runs.
Originally I just wrapped a TreeSet around sym2decl.values(), but
this slowed the output routine down -- there are a lot of symbols
and it had to allocate a tree node for each.  In the version I'm
committing here I've converted the set to a String[] to do the
sorting, to keep things fast.

I don't have any empirical evidence to prove that the speed difference
matters.  But I did it just the same. =)