Removed ./build as it was:  1) named incorrectly.  Should be something like 'build-rtj'.  2) placed incorrectly.  All scripts belong in the bin/ subdirectory.  3) Uses #!/bin/bash2 as the first line; bash2 is certainly not on     any debian machines, and I would guess is not on any machines     except the original committer's.  4) Doesn't support additional options, such as -Dfoo=bar, which     are likely to be necessary to enable various obscure flex features.     (The way bin/build-precisec and friends do it isn't ideal, but it's     better than nothing.)
I'm giving this the benefit of the doubt and assuming it's of general
use and not specific to the testing framework -- commands embedded
in the regression tests have no need to be short or user-friendly, since
no humans ever have to type them.

I've proactively removed this as I'm about to build a new distributions of
FLEX to upload to the web page for Christian Hammer, and I'm trying
to avoid confusing him with a script named 'build' that it may seem
to him that he should run.  It can be easily recovered with
  cd Harpoon/Code ; cvs update -r1.1 -p build
and then re-committed (if necessary) under a new name in a new location
with #3 (certainly) and #4 (hopefully) fixed.