As per Patrick's advice, create a 'SnapshotIterator' class which can be used when we want to mutate a collection underneath an iterator without the iterator noticing.  So the old:    Quad[] ql = (Quad[]) hc.getElements();    for (int i=0; i<ql.length; i++)        ql[i].accept(visitor); pattern can safely become:    for(Iterator<X> it=new SnapshotIterator(hc.getElementsI()); it.hasNext(); )        it.next().accept(visitor);
Note that this transformation is *not* safe without the construction of
the SnapshotIterator, as the iterator returned by hc.getElementsI() will
get *very* confused if your visitor starts modifying elements underneath
it.  The next commit will make the iterator "fail-fast" (see
http://java.sun.com/docs/books/tutorial/collections/implementations/general.html
) so that bugs of this sort are detected.

Also modified ReverseIterator so that it extends SnapshotIterator and does
not duplicate code.