REVERTED Wes' 23 Aug 2002 15:07:17 commit: POINTERS AND INTEGERS ARE NOT THE SAME THING.
The '%p' format specifier is to be used ONLY FOR POINTERS which may
very well be of a different size from integers.  ALL of the format
specifiers Wes changed from %08x to %p were INTEGERS not pointers,
and so the change introduced a BUG on platforms where
sizeof(int)!=sizeof(int*).