Be more aggressive in fixing up classes mentioned in other classes. This seems to prevent us from delaying the fixup until the "worst possible time", i.e. when FLEX is in the middle of changing class and method signatures during some transformation.
The specific bug I was seeing (that this commit fixes) had to do
with the unused and uncompiled class java.awt.RenderingHints
being "fixed up" in the middle of InitializerTransform when
HClass.getMutator() was called on java.awt.Component.  Component
had a *field* of type RenderingHints which got 'relinked' during
the creation of the mutator; this caused RenderingHints to be 'fixed up'
at a point *after* the split of Map.entrySet() to Map.entrySet$$initcheck().
Since RenderingHints was not only uninstantiated, but unreferenced in
compiled code (I think the Component class containing the RenderingHints
field was uninstantiable in this program), the MethodSplitter of course
did not split RenderingHints.entrySet() to RenderingHints.entrySet$$initcheck
(MethodSplitter didn't even know about RenderingHints because it didn't
show up anywhere in the ClassHierarchy).  Because RenderingHints didn't
contain the RenderingHints.entrySet$$initcheck method needed to implement
the (already split) Map.entrySet$$initcheck interface method,
AbstractClassFixupRelinker was dying on an assertion.  Changing to a
more aggressive strategy ensures that RenderingHints is fixed up
*before* InitializerTransform is run, bypassing the whole problem.